Increase number of PBKDF2 iterations
This solves issue #975 by rasing the number of iterations to be more
than 720,000. Also for the sake of simplicity all new backups will
use the same number of iterations insead of a random number in a
small range. This change is backwards compatable with the old backups
that had a random iteration count.